Refactor router structure
