+++ broadcaster/priv/static/admin/stream/stream.html
@@ -6,7 +6,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <link rel="icon" type="image/png" href="/favicon.png">
-  <link rel="stylesheet" type="text/css" href="stream/styles.css" />
+  <link rel="stylesheet" type="text/css" href="stream.css" />
   <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
   <title>Broadcaster | Elixir WebRTC</title>
 </head>
@@ -40,7 +40,7 @@
     </div>
   </div>
 
-  <script src="stream/stream.js"></script>
+  <script src="stream.js"></script>
 </body>
 
 </html>
