+++ broadcaster/priv/static/admin/stream/stream.js
