+++ broadcaster/priv/static/public/styles.css
@@ -25,3 +25,7 @@ body {
   max-height: 65%;
   border-radius: 8px;
 }
+
+.container {
+  padding: 10px;
+}
